Rename networking to network
