+++ src/pip/_internal/network/auth.py
